Test the gate rather than the flow it replaced
``guild_gate_test`` is what a community makes of a sign-in it did not serve:
who joins on arrival, who its narrowing counts, and what the gate says about
a session afterwards. The one that matters most is the rule being read live —
a community tightens its narrowing and the same session stops getting in,
without anybody signing in again.

``guild_login_flow_test`` becomes ``provider_login_session_test``. What it was
really testing is what one sign-in does to the session behind it — the step-up
union, each provider's own account of its own event, the revoked predecessor —
and none of that was ever about which community somebody was heading for. The
guild-route half goes with the routes.

The agreement matrix keeps its job and changes its axes: the second dimension
was which communities a session had signed into, and is now what each provider
asserted. Rule shapes times standings, both layers, still to the row.

Three things the tests found:

* a community at capacity rolled back the transaction the sign-in was riding
  on. Each join takes a savepoint now, so one full community undoes its own
  attempt and nothing else.
* ``SessionAssurance`` dropped the assertions on the way through the token,
  because a model only keeps the fields it declares.
* the JSONB parameter was being handed an already-encoded string, which
  arrives as a JSON scalar rather than an object to index into.

``create_guild_auth_policy`` connects the community to the provider first. The
endpoint refuses a requirement naming a provider the community does not count
as its own, so a policy without a connection was a state production could not
reach and the factory should not have been able to build.

+++ backend/app/db/authorization.py
@@ -56,7 +56,7 @@ __all__ = [
 #: the credential proved; ``p_claims`` is what those providers asserted,
 #: ``{"12": {"hd": ["acme.com"]}}``.
 GUILD_CONNECTION_ADMITS = """\
-CREATE OR REPLACE FUNCTION public.guild_connection_admits(p_guild_id integer, p_providers integer[], p_claims jsonb, p_provider_id integer DEFAULT NULL)
+CREATE OR REPLACE FUNCTION public.guild_connection_admits(p_guild_id integer, p_providers integer[], p_claims jsonb, p_provider_id integer DEFAULT NULL::integer)
  RETURNS boolean
  LANGUAGE sql
  STABLE
@@ -99,7 +99,7 @@ $function$
 #: Gate 0b: the same question for the current request, off the GUCs the
 #: session context sets. This is what the policy legs call.
 GUILD_CONNECTION_SATISFIED = """\
-CREATE OR REPLACE FUNCTION public.guild_connection_satisfied(p_guild_id integer, p_provider_id integer DEFAULT NULL)
+CREATE OR REPLACE FUNCTION public.guild_connection_satisfied(p_guild_id integer, p_provider_id integer DEFAULT NULL::integer)
  RETURNS boolean
  LANGUAGE sql
  STABLE
